Fixed 30052. Cleanup all 'Unnecessary cast' warnings from Eclipse.
As those extra casts are not useful and make the code harder to read/understand
-> Remove all those

--HG--
extra : rebase_source : 63fb38c1cbd7435bf21e9bd4f7369c07c1f026b9

+++ modules/org.openbravo.advpaymentmngt/src/org/openbravo/advpaymentmngt/actionHandler/PaymentProposalPickEditLines.java
@@ -121,7 +121,7 @@ public class PaymentProposalPickEditLines extends BaseProcessActionHandler {
     int cont = 0;
     String differentPaymentMethod = "false";
     for (int i = 0; i < selectedLines.length(); i++) {
-      JSONObject selectedLine = selectedLines.getJSONObject((int) i);
+      JSONObject selectedLine = selectedLines.getJSONObject(i);
       log.debug(selectedLine);
       BigDecimal paidAmount = new BigDecimal(selectedLine.getString("payment"));
